requirements.txt
google-api-python-client==2.114.0
google-auth-oauthlib==1.2.0
numpy==1.26.3
pandas==2.1.4
